Update main.tf to make acr name globally unique
Using the hard coded cyberanalyzeracr name throws the following error:

Error: the name "cyberanalyzeracr" used for the Container Registry needs to be globally unique and isn't available: The registry cyberanalyzeracr is already in use.

The name is in use by our great teacher, Ed Donner. Adding a randomly generated suffix fixes the issue.